Clean UDevEvent class and rename file to udev_event.*
